requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.12.2
branca==0.6.0
capmonster-python==2.2
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
colorama==0.4.6
comtypes==1.1.14
contourpy==1.0.6
customtkinter==5.1.2
cycler==0.11.0
darkdetect==0.8.0
decorator==4.4.2
distlib==0.3.6
docutils==0.19
exceptiongroup==1.0.0rc9
filelock==3.8.0
folium==0.14.0
fonttools==4.38.0
frozenlist==1.3.3
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
imageio==2.27.0
imageio-ffmpeg==0.4.8
Jinja2==3.1.2
Kivy==2.1.0
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.4.5
Kivy-Garden==0.1.5
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.6.2
moviepy==1.0.3
multidict==6.0.4
numpy==1.24.1
openai==0.26.1
opencv-python==4.7.0.68
outcome==1.2.0
packaging==23.0
pandas==1.5.2
Pillow==9.4.0
pipenv==2022.11.11
platformdirs==2.5.4
plotly==5.13.0
proglog==0.1.10
pyaes==1.6.1
pyasn1==0.4.8
PyAudio==0.2.13
pycparser==2.21
pydub==0.25.1
pygame==2.1.2
Pygments==2.14.0
pyparsing==3.0.9
pypiwin32==223
PySocks==1.7.1
python-dateutil==2.8.2
pyttsx3==2.90
pytz==2022.7
pywin32==305
requests==2.28.1
rfc3986==1.5.0
rsa==4.9
selenium==3.141.0
six==1.16.0
sklearn==0.0.post1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SpeechRecognition==3.9.0
Telethon==1.27.0
tenacity==8.1.0
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
virtualenv==20.16.7
virtualenv-clone==0.5.7
wsproto==1.2.0
yarl==1.8.2
